Introduce new `schema::system` module
