+++ p2panda-rs/src/schema/system/mod.rs
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+//! System schemas are p2panda's built-in schema type.
+//!
+//! They are defined as part of the p2panda specification and may have differ from application
+//! schemas in how they are materialised.
+
+mod error;
+mod schema;
+
+pub use error::SystemSchemaError;
+pub use schema::{FieldType, SchemaFieldView, SchemaView};
